comp4350group4 - Final/requirements.txt
Babel==0.9.6
Flask==0.9
Flask-Babel==0.8
Flask-Login==0.1.3
Flask-Mail==0.7.4
Flask-OpenID==1.1.1
Flask-SQLAlchemy==0.16
Flask-WTF==0.8.2
Jinja2==2.6
SQLAlchemy==0.7.9
Tempita==0.5.1
WTForms==1.0.2
WTForms-JSON==0.1.2
Werkzeug==0.8.3
argparse==1.2.1
blinker==1.2
decorator==3.4.0
distribute==0.6.24
## FIXME: could not find svn URL in dependency_links for this package:
flup==1.0.3.dev-20110405
python-openid==2.2.5
pytz==2012j
speaklater==1.3
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
